+++ challenge_prex/src/response_dto.rs
@@ -2,9 +2,9 @@ use serde::{Serialize, Deserialize};
 use chrono::NaiveDate;
 use rust_decimal::Decimal;
 
-#[derive(Debug, Serialize,Deserialize)]
+#[derive(Default,Debug, Serialize,Deserialize)]
 pub struct ClientBalance {
-    pub client_id: u128,
+    pub client_id: u32,
     pub client_name: String,
     pub bith_date: NaiveDate,
     pub document_number: String,
